CHROMIUM: cros-ec: Only update from known PD device RW firmware
If a PD device is currently operating in RW and the RW image is not
known as an old image, we don't want to perform an update -- doing so
might mean rolling back newer firmware. Achieve this by maintaining a
list of known RW hashes from which we wish to upgrade.

BUG=chrome-os-partner:34599
TEST=Manual on Samus. Insert Zinger with latest firmware, verify that
firmware isn't re-flashed. Corrupt RW firmware on Zinger, verify that FW
is reflashed successfully. Roll back to zinger_000003, verify that FW is
reflashed to latest. Flash a new valid unknown RW image on Zinger,
verify that FW is not reflashed.

Change-Id: Ibd280b145d515cc91b60a433aefa536270da5d85
Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/236759
Reviewed-by: Alec Berg <alecaberg@chromium.org>