CHROMIUM: HID: i2c-hid - check if device is there before really probing
Unfortunately vendors like to interchange touchpad parts between factory
runs, but keep the same DTS, so we can't be quite sure that the device
is actually present in the system even if it is described in the DTS.

Before trying to properly initialize the touchpad iand generate bunch of
errors let's first see it there is anything at the given address. If we
get error fail silently with -ENXIO.

BUG=chrome-os-partner:34495
TEST=build and boot unit with/without Synaptics trackpad, verify the
system functions properly.

Change-Id: I053aece217dfbba8aac223fa7ca010fec9ccdaed
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/236896
Reviewed-by: Andrew Duggan <aduggan@synaptics.com>
Reviewed-by: Benson Leung <bleung@chromium.org>
Conflicts:
	drivers/hid/i2c-hid/i2c-hid.c
[rebase412(groeck): Context conflicts]
Signed-off-by: Guenter Roeck <groeck@chromium.org>