iommu: move pgsize_bitmap from struct iommu_ops to struct iommu_domain
struct iommu_ops contains function pointers for IOMMU driver callbacks
in order to implement the core IOMMU API. Amongst these pointers is
an unsigned long pgsize_bitmap field, which is problematic because the
set of supported page sizes is not necessarily the same across all
instances of a given IOMMU type in the system. Furthermore, the set of
supported page sizes may be restricted following domain initialisation
when a particular page table format is chosen for the domain.

This patch moves the field from iommu_ops into the iommu_domain and
updates all users accordingly.

Acked-by: Daniel Kurtz <djkurtz@chromium.org>
Acked-by: Thierry Reding <treding@nvidia.com>
Acked-by: Marek Szyprowski <m.szyprowski@samsung.com>
Acked-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>