video: tegra: nvmap: don't pass seq_write() failures back to seq_file
seq_write() returns -1 when the buffer is full.  If the seq_write()
error is ignored, seq_file will eventually check for this and try again
with a larger buffer.

The show() op can use this as a cue to stop writing, but it should
return 0 anyway.  If it returns -1 back to seq_file, this is treated as
a "real" error and interpreted by userspace as EPERM.

Bug: 19271939

Change-Id: I1b06b92199d7d52919f8e1930d7b2c8d8ae90718
Signed-off-by: Greg Hackmann <ghackmann@google.com>