UPSTREAM: clk: qcom: Introduce parent_map tables
In the current parent mapping code, we can get duplicate or inconsistent
indexes, which leads to discrepancy between the number of elements in the
array and the number of parents. Until now, this was solved with some
reordering but this is not always possible.

This patch introduces index tables that are used to define the relations
between the PLL source and the hardware mux configuration value.
To accomplish this, here we do the following:
 - Define a parent_map struct to map the relations between PLL source index
 and register configuration value.
 - Add a qcom_find_src_index() function for finding the index of a clock
 matching the specific PLL configuration.
 - Update the {set,get}_parent RCG functions use the newly introduced
 parent_map struct.
 - Convert all existing drivers to the new parent_map tables.

Signed-off-by: Georgi Djakov <georgi.djakov@linaro.org>
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
(cherry picked from commit 293d2e97b37f545bb36aef78cd549d9e6cd66e7f)

BUG=chrome-os-partner:49255
TEST=Boots to prompt over serial with ramdisk

Change-Id: I759d8298df0b8599669947cb1621e5dee462fd9d
Signed-off-by: Matthew McClintock <mmcclint@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/331709
Commit-Ready: Dmitry Torokhov <dtor@chromium.org>
Tested-by: Matthew McClintock <mmcclint@qca.qualcomm.com>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>