UPSTREAM: tytso: ext4 crypto: add ext4 encryption facilities
On encrypt, we will re-assign the buffer_heads to point to a bounce
page rather than the control_page (which is the original page to write
that contains the plaintext). The block I/O occurs against the bounce
page.  On write completion, we re-assign the buffer_heads to the
original plaintext page.

On decrypt, we will attach a read completion callback to the bio
struct. This read completion will decrypt the read contents in-place
prior to setting the page up-to-date.

The current encryption mode, AES-256-XTS, lacks cryptographic
integrity. AES-256-GCM is in-plan, but we will need to devise a
mechanism for handling the integrity data.

Signed-off-by: Michael Halcrow <mhalcrow@google.com>
Signed-off-by: Ildar Muslukhov <ildarm@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
(cherry picked from commit b30ab0e03407d2aa2d9316cba199c757e4bfc8ad)

BUG=chromium:347322
TEST=Compiled, Installed on samus
TEST=Test with e2fsprogs 1.43-WIP

Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Change-Id: I8d784936ecc2d69056d9b69778e52f7f8285953c
Reviewed-on: https://chromium-review.googlesource.com/321670
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>