arm: perf: treat PMUs as CPU affine
In multi-cluster systems, the PMUs can be different across clusters, and
so our logical PMU may not be able to schedule events on all CPUs.

This patch adds a cpumask to encode which CPUs a PMU driver supports
controlling events for, and limits the driver to scheduling events on
those CPUs, and enabling and disabling the physical PMUs on those CPUs.
The cpumask is built based on the interrupt-affinity property, and in
the absence of such a property a homogenous system is assumed.

Acked-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>
(cherry picked from commit cc88116da0d18b8292f5437dbc0c4683c8a34ac1)
Signed-off-by: Alex Shi <alex.shi@linaro.org>
(cherry picked from commit f7a203e1f2fea8a2fcc4828551d6caae8521d3d9)
Signed-off-by: Alex Shi <alex.shi@linaro.org>