CHROMIUM: mfd: cros_ec: add debugfs, console log file
If the EC supports the new CONSOLE_READ command type, then we
place a console_log file in debugfs for that EC device which allows
us to grab EC logs. The kernel will poll every 30 seconds for the
log and keep its own buffer, but userspace should grab this and
write it out to some logs which actually get rotated.

[drinkcat: Cherry-picked from 3.18 tree,
commit 0996ea6a5852 ("CHROMIUM: mfd: cros_ec: add debugfs, console log file")

Locations of files changed, cros_ec_commands need to be allocated
with kmalloc. Fixed log_work: No need to set the whole ec_buffer
to zero.]

BUG=chromium:492721
BUG=chromium:527904
TEST=on kevin, cat /sys/kernel/debug/cros_ec/console_log and see
  that it has data, cat again and see that it blocks until the
  kernel has polled the EC and data is available

Change-Id: I4effa3febec1c90661e0b1ca76a4e21524c987af
Signed-off-by: Eric Caruso <ejcaruso@chromium.org>
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/372758
Reviewed-by: Guenter Roeck <groeck@chromium.org>

Conflicts:
	drivers/platform/chrome/Makefile

[rebase49(groeck): Fix context conflicts;
		cros_ec_dev_platform -> cros_ec_platform]
Signed-off-by: Guenter Roeck <groeck@chromium.org>