iommu: Make sure a device is always attached to a domain
Make use of the default domain and re-attach a device to it
when it is detached from another domain. Also enforce that a
device has to be in the default domain before it can be
attached to a different domain.

Signed-off-by: Joerg Roedel <jroedel@suse.de>
(cherry picked from commit e39cb8a3aa988a74433a3f26443b454cca033651)
Signed-off-by: Alex Shi <alex.shi@linaro.org>