CHROMIUM: [media] v4l: fix VIDIOC_DQEVENT in 64- / 32-bit compatibility mode
Per ABI, the alignment of a structure or a union is the maximum
alignment of its members. If we don't specify each member in the
union explicitly in v4l2_event32 like we do in v4l2_event, the
union in v4l2_event32 will not be aligned to the maximum alignment
of its members. The size of v4l2_event32 will be different from
32-bit v4l2_event and VIDIOC_DQEVENT32 will not match 32-bit
VIDIOC_DQEVENT.

BUG=chrome-os-partner:38980
TEST=wrote a 32-bit test program to run on 64-bit kernel

Change-Id: I892a5512d7a196518e7e6fc14ef4b9dc390d224a
Signed-off-by: Wu-Cheng Li <wuchengli@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/275044
Reviewed-by: Pawel Osciak <posciak@chromium.org>

Conflicts:
	drivers/media/v4l2-core/v4l2-compat-ioctl32.c

[rebase44(groeck): Resolved conflicts; updated subject line]
Signed-off-by: Guenter Roeck <groeck@chromium.org>