Input: sensors: add documentation for sensor devicetree node
The ltr559.txt is used for explaining the meaning of the ltr559 driver's
devicetree parameters, which the ltr559 is a light and proximity sensor.
The kionix_accel.txt is used for explaining the meaning of the kionix_accel
driver's devicetree parameters, which the kionix_accel is a 3-axis
accelerometer sensor.