arch/arm: fix ARM_ip overwrite in syscall_trace()
When tracing system calls using ptrace(PTRACE_SYSCALL, ...),
the tracee will crash with a SIGSEGV and instruction pointer set to zero.

ARM_ip is used to indicate syscall entry/exit and restored to its original
content after the tracer has been notified.

This patch removes the code that overwrites the variable
that contains the original content of ARM_ip, which leads to an invalid
ARM_ip content and therefore a SIGSEGV in the tracee.

Bug: 178583

Change-Id: I6d735b39b2dd976ad767fbc226a999760b363696
Signed-off-by: Julian Hector <julian.r.hector@gmail.com>